+++ examples/resources/nxos_default_qos_class_map/import.sh
@@ -1,0 +1,1 @@
+terraform import nxos_default_qos_class_map.example "sys/ipqos/dflt/c/name-[Voice]"