+++ examples/resources/nxos_default_qos_class_map/resource.tf
@@ -1,0 +1,4 @@
+resource "nxos_default_qos_class_map" "example" {
+  name       = "Voice"
+  match_type = "match-any"
+}